Fixed: binary not installed Removed unused variables and fixed some other compiler warnings
git-svn-id: svn://svn.code.sf.net/p/keepassx/code/trunk@163 b624d157-de02-0410-bad0-e51aec6abb33

+++ src/lib/WaitAnimationWidget.cpp
@@ -55,7 +55,7 @@ void WaitAnimationWidget::stop(){
 
 void WaitAnimationWidget::setRefreshRate(int fps){
 	DiffAngle=1.0f/((60.0f/(float)speed)*(float)fps);
-	timer.setInterval((1.0f/(float)fps)*1000.0f);
+	timer.setInterval((int) ((1.0f/(float)fps)*1000.0f));
 }
 
 void WaitAnimationWidget::refreshAnimation(){
